Make lglob not crash when it's built with MSVC x64
The return type of _findfirst is intptr_t. When the target is x64,
intptr_t is 64-bit, while long is 32-bit, so using long truncates
the result.

+++ lglob.h
@@ -62,7 +62,7 @@
 					char dir[_MAX_DIR];	\
 					char fname[_MAX_FNAME];	\
 					char ext[_MAX_EXT];	\
-					long handle;
+					intptr_t handle;
 
 #else
 #if MSDOS_COMPILER==WIN32C && !defined(_MSC_VER) /* Borland C for Windows */
